Refactor: split ThreadLocal into module
